Working tech tree and made the fist two recipes. Attached the tech tree to the crafting system and made it networking.

+++ Assets/scripts/ItemScripts/AxeScript.cs
@@ -6,7 +6,7 @@ using Mirror;
 public class AxeScript : GameItem
 {
     public string axeType;
-    public double power;
+    public int power;
 
     public override void interact(PlayerControl player)
     {
